Aplicando validação de Id para que a apalicação não caia ao produrar um id que não existe.

+++ src/main/java/br/com/digitalhouse/oficina/resource/ClienteResource.java
@@ -2,6 +2,7 @@ package br.com.digitalhouse.oficina.resource;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import br.com.digitalhouse.oficina.repository.ClienteRepository;
 
 import br.com.digitalhouse.oficina.dto.ClienteInsertDTO;
 import br.com.digitalhouse.oficina.model.Cliente;
@@ -25,6 +27,9 @@ public class ClienteResource {
 	
 	@Autowired
 	private ClienteService clienteService;
+	
+	@Autowired
+	private ClienteRepository clienteRepository;
 
 //	@Autowired
 //	public ClienteResource(ClienteService clienteService) {
@@ -50,6 +55,10 @@ public class ClienteResource {
 	
 	@PutMapping("/{id}")
 	public ResponseEntity<Void> update(@PathVariable Long id, @RequestBody ClienteInsertDTO clienteDTO){
+		if (!clienteRepository.existsById(id)) {
+			return ResponseEntity.notFound().build();
+		}
+		
 		Cliente cliente = new Cliente();		
 		cliente.setId(id);
 		cliente.setNome(clienteDTO.getNome());
@@ -64,9 +73,14 @@ public class ClienteResource {
 	@GetMapping("/{id}")  // /clientes/3
 	public ResponseEntity<Cliente> findById(@PathVariable Long id){
 		
-		Cliente cliente = this.clienteService.findById(id);
 		
-		return ResponseEntity.ok(cliente);
+		Optional<Cliente> cliente = clienteRepository.findById(id);
+
+		if (cliente.isPresent()) {
+			return ResponseEntity.ok(cliente.get());
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 	
 	@GetMapping // /clientes
@@ -81,6 +95,10 @@ public class ClienteResource {
 	// Funcionando somente para clientes sem veiculos cadastrados
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> delete(@PathVariable Long id){
+		
+		if (!clienteRepository.existsById(id)) {
+			return ResponseEntity.notFound().build();
+		}
 		
 		this.clienteService.deleteById(id);
 		return ResponseEntity.noContent().build();
